backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.6.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
sentence-transformers==2.5.1
torch==2.2.2
pgvector==0.2.5
pytesseract==0.3.10
pdfminer.six==20231228
python-docx==1.1.2
spacy==3.7.2
